### 2018-01-08 #525 @arfon
@rreinecke - thanks for your submission to JOSS.

There are a few issues with your submission as it currently stands:

- The source code doesn't seem to be hosted on a platform such as GitHub/Bitbucket/Sourceforge which is going to make it very difficult for your reviewers to open flag issues that arise during review with your project.
- I don't seem to be able to find a `paper.md` file associated with your submission. Please review [our submission guidelines](http://joss.theoj.org/about#paper_structure) on what this needs to contain.
- I don't seem to be able to find an open source license associated with your submission. All submissions to JOSS must have a clear statement on software licensing and use of the [OSI approved](https://opensource.org/licenses/alphabetical) licenses.

Let me know if you have any questions.
