### 2018-01-08 #525 @rreinecke
Hi Arfon,

The code is hosted on github:
https://github.com/rreinecke/global-gradient-based-groundwater-model
and should contain all the required files and is liked on the submited
homepage which is also hosted on github.
I'm sorry but the submission form was unclear on that topic and I've
submitted the official url. Maybe you can add a hint that it is
automatically processed and that one should use the github url of the
repository.

Cheers
Robert.

Quoting Arfon Smith <notifications@github.com>:

> @rreinecke - thanks for your submission to JOSS.
>
> There are a few issues with your submission as it currently stands:
>
> - The source code doesn't seem to be hosted on a platform such as
> GitHub/Bitbucket/Sourceforge which is going to make it very
> difficult for your reviewers to open flag issues that arise during
> review with your project.
> - I don't seem to be able to find a `paper.md` file associated with
> your submission. Please review [our submission
> guidelines](http://joss.theoj.org/about#paper_structure) on what
> this needs to contain.
> - I don't seem to be able to find an open source license associated
> with your submission. All submissions to JOSS must have a clear
> statement on software licensing and use of the [OSI
> approved](https://opensource.org/licenses/alphabetical) licenses.
>
> Let me know if you have any questions.
>
> --
> You are receiving this because you were mentioned.
> Reply to this email directly or view it on GitHub:
> https://github.com/openjournals/joss-reviews/issues/525#issuecomment-356002600
