### 2018-01-26 #548 @whedon
[REVIEW]: G³M-f a global gradient-based groundwater modelling framwork
**Submitting author:** @rreinecke (<a href="http://orcid.org/0000-0001-5699-8584">Robert Reinecke</a>)
**Repository:** <a href="https://github.com/rreinecke/global-gradient-based-groundwater-model" target ="_blank">https://github.com/rreinecke/global-gradient-based-groundwater-model</a>
**Version:** v1.0.0
