### 2018-01-26 #548 @Kevin-Mattheus-Moerman
@rreinecke can you change the name of your paper file to use small letters (rather than capital letters). I think this might be why it isn't finding it.
