### 2018-01-26 #548 @Kevin-Mattheus-Moerman
@rreinecke That worked. Btw you can also ask whedon to generate the PDF yourself. I noticed the are several DOI's missing for the reference. Can you add them?
