### 2018-02-08 #548 @wkearn
Here is my review!

G³M seems like a powerful groundwater modeling framework with a lot of potential for becoming a reusable component in hydrological models at a variety of scales. My main concern is that under the current state of its documentation, it would be very challenging for a new user to start running their own models or to couple G³M to other models.

I also have a few minor concerns about the installation process, some of which might not be bugs in G³M, but in dependencies.

# Documentation

## Governing equations, etc.

I would like to see a summary of the governing equations, the boundary conditions, and the numerical methods behind this model. This would be really helpful in comparing it to other models (say MODFLOW).

## The simple model example

A walkthrough of the simple model would be a great help in understanding exactly what a user would need to construct their own model. How would a hydrologist go from 0 to output that they can analyze? I think you have the building blocks of this in your current documentation (such as the configuration file example), but some more explanation of what each of the options controls and how to implement the `GW_Interface` would go a long way towards making it easier to get started.

## A more complex example

It would also be good to see a more complex example that exercises more of the software options, especially coupling G³M to other models, since that seems to be a key argument for G³M.

# Installation

Dependencies are not automatically installed, which is particularly tricky for Google Test, which now recommends including libgtest in the project, rather than relying on a precompiled shared library (so that there isn't a libgtest package in apt, for instance). I built libgtest on my own and then copied it into `/usr/local/lib`, which is a complicated thing for a new user to be asked to do without further instruction. This could be resolved by either distributing Google Test with the project, by providing instructions for the user to put the Google Test code into the project themselves (setting appropriate environment variables) or by modifying the build process so that a user can build the models without building the tests.

I am having trouble compiling the tests because of what might be a problem with my Boost installation. I can comment out the offending lines in `tagged_integer.hpp`, and the tests compile and run properly. I am more concerned that this causes the documented build process to fail.

This is the error that I get. If this is a Boost bug, I'd be happy to run it down with the Boost developers. I can also open an issue on the G³M repository if you would like to handle it over there.

```
[ 95%] Building CXX object CMakeFiles/runUnitTests.dir/tests/testrunner.cpp.o
In file included from /usr/include/boost/log/sinks/attribute_mapping.hpp:23:0,
                 from /usr/include/boost/log/sinks/syslog_backend.hpp:32,
                 from /usr/include/boost/log/sinks.hpp:34,
                 from /home/wkearn/local/src/global-gradient-based-groundwater-model/tests/Unit/Model/../../../src/Model/../Misc/../Logging/Sinks.hpp:20,
                 from /home/wkearn/local/src/global-gradient-based-groundwater-model/tests/Unit/Model/../../../src/Model/../Misc/../Logging/Logging.hpp:17,
                 from /home/wkearn/local/src/global-gradient-based-groundwater-model/tests/Unit/Model/../../../src/Model/../Misc/Helpers.hpp:26,
                 from /home/wkearn/local/src/global-gradient-based-groundwater-model/tests/Unit/Model/../../../src/Model/ExternalFlows.hpp:28,
                 from /home/wkearn/local/src/global-gradient-based-groundwater-model/tests/Unit/Model/../../../src/Model/Node.hpp:34,
                 from /home/wkearn/local/src/global-gradient-based-groundwater-model/tests/Unit/Model/Node__TEST.hpp:2,
                 from /home/wkearn/local/src/global-gradient-based-groundwater-model/tests/Unit/UnitTests.hpp:2,
                 from /home/wkearn/local/src/global-gradient-based-groundwater-model/tests/testrunner.cpp:2:
/usr/include/boost/log/detail/tagged_integer.hpp: In function ‘bool boost::log::v2_mt_posix::aux::operator<(const boost::log::v2_mt_posix::aux::tagged_integer<IntT, TagT>&, const boost::log::v2_mt_posix::aux::tagged_integer<IntT, TagT>&)’:
/usr/include/boost/log/detail/tagged_integer.hpp:96:1: error: parse error in template argument list
 BOOST_LOG_TAGGED_INTEGER_OP(<)
 ^
CMakeFiles/runUnitTests.dir/build.make:254: recipe for target 'CMakeFiles/runUnitTests.dir/tests/testrunner.cpp.o' failed
make[2]: *** [CMakeFiles/runUnitTests.dir/tests/testrunner.cpp.o] Error 1
CMakeFiles/Makefile2:104: recipe for target 'CMakeFiles/runUnitTests.dir/all' failed
make[1]: *** [CMakeFiles/runUnitTests.dir/all] Error 2
Makefile:94: recipe for target 'all' failed
make: *** [all] Error 2
```
