### 2018-02-09 #548 @rreinecke
@wkearn Thank you very much for your very helpful feedback! @Kevin-Mattheus-Moerman yes I will approach the mentioned issues  in the next couple days.
