### 2018-02-10 7106683
added google test to cmake build
