### 2018-02-10 0638fdf
removed unused logging code to track down boost error
