+++ src/Logging/Sinks.hpp
@@ -36,29 +36,6 @@ typedef sinks::synchronous_sink<sinks::text_ostream_backend> text_sink;
 /**
  * @deprecated Use methods from DataProcessing
  */
-class InfoSinkInterface :
-        public sinks::basic_sink_backend<
-                sinks::combine_requirements<
-                        sinks::synchronized_feeding,
-                        sinks::flushing
-                >::type
-        > {
-    private:
-        // The file to write the collected information to
-        std::ofstream m_file;
-    public:
-        InfoSinkInterface(std::string file_name) {};
-
-        virtual ~InfoSinkInterface() {};
-
-        // The function consumes the log records that come from the frontend
-        virtual void consume(logboost::record_view const &rec) = 0;
-
-        void flush() {};
-
-        void write_data() {};
-};
-
 /*class CSVWriter : public InfoSinkInterface {
 
 };
