### 2018-02-10 #548 @rreinecke
@wkearn concerning the libboost error. I can reproduce it with version 1.58 but not with 1.54 and I have currently no idea on what the issue actually is. Would you be so kind and open it as an issue that the build fails with that specific version and if you got the time run it down with the boost mailing list?

I've added the gtest framework now to the project build - no more need for a manual install
