### 2018-02-10 623433e
added equation and boundary condition explenation
