### 2018-02-10 f19e8c1
fixing equation image
