### 2018-02-14 254394f
Update README.md
Updated config description
