### 2018-02-14 4da8935
Update README.md
added text for model coupling
