+++ README.md
@@ -207,7 +207,16 @@ simple_model
 It will yield a depth to water table CSV file called wtd.csv for a simple model.
 
 ## Deployment in other models
-Just implement the GW_interface and provide a DataReader.
+The main steps towards your own model is to implement the GW_interface and provide a DataReader.
+A standlone version can be easily implemented by extending the simple example provided above.
+
+### In memory coupling
+G³M-f is written with the coupling to other models in mind.
+In contrast to other model coupling efforts, it is not necessary to write out files in one model and read them in in another model.
+You can diretly link G³M-f with your existing executable and by providing a class in your already existing model code that implements the gw_interface, you are free to call the simulate() function at any timestep you like.
+Furthermore, the interface provides pointer containers and callbacks to transfer data in memory without the need to waste time on I/O.
+
+Please contact us if you need advice.
 
 ## Running the tests
 Automated tests consits of gunit test which are compiled automatically with the attached cmake file.
