### 2018-02-14 ec12f27
Update README.md
path to example implementation
