### 2018-02-14 7dc9575
Update README.md
added link to modflow documentation
