### 2018-02-14 bccf712
Update README.md
added input data requirements for the simple model
