### 2018-02-14 #548 @rreinecke
@wkearn

Addressed issues:
(R1): Governing equations: I've added additional descriptions concerning the equations. In general they are very similar to MODFLOW why I felt reluctant to repeat their documentation effort. I've added a link to the documentation in the description text
(R2): The simple model example: I've extended the simple model explanation - hope it is now more streamlined and helps as a basis for the user to construct his/her own model. Also added description of main model configuration parameters.
(R3): A more complex example: To be honest I've pondered on this one quite a while. Of course I have a more complex model, which is my global groundwater model, and I will link the publication as soon as it is published - but for now I probably don't have the time to provide a well documented complex example. I've added a description on the coupling process and a note that developers are welcome to contact me on coupling efforts.
(R4): Installation: As noted in the post above - 1) gtest is now shipped with the code 2) the libboost bug is reproducible with a specific version, but still a mystery to me.

Cheers
Robert.
