### 2018-02-18 #548 @wkearn
@rreinecke This looks great!

I am happy with a link to Modflow for documentation. As you say, their documentation is a lot more comprehensive, and if G³M implements basically the same equations, there is no point in duplicating that.

The simple model walkthrough is now more clear. I would encourage you to keep updating the documentation as you and others develop more applications with G³M, including your global groundwater model. Looking forward to seeing that publication when it comes out!

The installation works now, with the exception of the libboost bug. I haven't had time yet to track this down, but I will let you know if I hear anything about it.

@Kevin-Mattheus-Moerman I'm happy to recommend this for acceptance.
