### 2018-02-19 #548 @Kevin-Mattheus-Moerman
Thanks @wkearn! 🤖 Are you done with the tickboxes at the top of this issue?

@rreinecke can you make a DOI to the archived version of the reviewed software available? This is needed to process acceptance of this submission.
