### 2018-02-19 #548 @rreinecke
@wkearn Thank you very much for your helpful feedback!
I'm very much looking forward to share the global model results with you and the research community.

@Kevin-Mattheus-Moerman https://zenodo.org/badge/latestdoi/109667597
