### 2018-02-19 b08f8ff
Update README.md
added JOSS badge
